Fix rusqlite 0.40 breaking change: usize/u64 no longer implement ToSql/FromSql
rusqlite 0.40 removed ToSql/FromSql for usize/u64 (platform-width
ambiguity - SQLite's INTEGER affinity is a signed 64-bit value, so i64
is the only honest representation). Four real call sites broke:

- feedback.rs: edit_distance (Option<usize>) bound directly as a query
  param, and read back the same way.
- feedback.rs: `limit: usize` bound directly as a LIMIT param.
- feedback.rs: a Vec<usize> read back directly from a query.
- tracker.rs: CompensationBand's min/max/median (u64) bound directly
  as insert params.

Fixed by casting to i64 at the SQL boundary in each case (matching the
`limit as i64` pattern already used elsewhere in this file) and back to
usize/u64 on read where the in-memory type should stay unsigned. Column
types are unchanged (SQLite INTEGER was always the real storage type
here regardless of what Rust type got bound to it).

Verified: clippy/fmt/build clean, all 71 tests pass including
test_feedback_tracker_full_cycle, which exercises this exact path.

+++ src/pipeline/tracker.rs
@@ -762,9 +762,12 @@ impl PipelineTracker {
                 role.fit_score,
                 serde_json::to_string(&role.market_demand).unwrap_or_default(),
                 role.compensation_band.currency,
-                role.compensation_band.min,
-                role.compensation_band.max,
-                role.compensation_band.median,
+                // rusqlite's ToSql no longer covers u64 directly (removed
+                // upstream over platform-width ambiguity) - i64 is what
+                // SQLite's INTEGER affinity actually stores.
+                role.compensation_band.min as i64,
+                role.compensation_band.max as i64,
+                role.compensation_band.median as i64,
                 role.compensation_band.source,
                 serde_json::to_string(&role.typical_requirements).unwrap_or_default(),
                 serde_json::to_string(&role.top_companies).unwrap_or_default(),
